+++ lib/jdbc-helper/connector.rb
@@ -16,4 +16,5 @@ end#JDBCHelper
 
 require 'jdbc-helper/connector/oracle_connector'
 require 'jdbc-helper/connector/mysql_connector'
+require 'jdbc-helper/connector/postgres_connector'