+++ test/test_sql.rb
@@ -1,5 +1,6 @@
 require 'helper'
-include JDBCHelper
+SQL = JDBCHelper::SQL
+SQLPrepared = JDBCHelper::SQLPrepared
 
 # WARNING: tests assumes ordered hash
 class TestSQL < Test::Unit::TestCase
@@ -17,7 +18,7 @@ class TestSQL < Test::Unit::TestCase
         SQL.value(BigDecimal.new("0.00000000000000000000009999999999999999999"))
     assert_equal "'sysdate'",         SQL.value('sysdate')
     assert_equal "'A''s'",            SQL.value("A's")
-    assert_equal "sysdate",           SQL.value(JDBCHelper::SQL('sysdate'))
+    assert_equal "sysdate",           SQL.value(SQL.expr('sysdate'))
 
     assert_raise(NotImplementedError) { SQL.value(Time.now) }
   end
@@ -36,6 +37,26 @@ class TestSQL < Test::Unit::TestCase
     assert_raise(ArgumentError) { SQL.order(:a, :b, "c 'd") }
   end
 
+  def test_complex_table_wrapper_query
+    assert_equal "select a + b sum from test where a = 1 and b >= 1 and b <= 10 and c in ('a', 'b', 'c') and d = sysdate and e is not null and f > 100 and g < 100 and h like 'ABC%' and i not like 'ABC%' and j <= sysdate order by a + b desc",
+      JDBCHelper::TableWrapper.new(nil, :test).where(
+        :a => 1,
+        :b => (1..10),
+        :c => %w[a b c],
+        :d => SQL.expr('sysdate'),
+        :e => SQL.not_null,
+        :f => SQL.gt(100),
+        :g => SQL.lt(100),
+        :h => SQL.like('ABC%'),
+        :i => SQL.not_like('ABC%'),
+        :j => SQL.le( SQL.expr('sysdate') )
+      ).select(
+        SQL.expr('a + b sum')
+      ).order(
+        SQL.expr('a + b desc')
+      ).sql
+  end
+
   def test_where
     assert_equal "where a = 1",                   SQL.where(:a => 1)
     assert_equal "where a = 12345678901234567890.1234567890123456789", 
@@ -48,13 +69,14 @@ class TestSQL < Test::Unit::TestCase
     assert_equal "where a is null",               SQL.where(:a => nil)
     assert_equal "where a is not null",           SQL.where(:a => SQL.not_nil)
     assert_equal "where a is not null",           SQL.where(:a => SQL.not_null)
-    assert_equal "where a = sysdate",             SQL.where(:a => JDBCHelper::SQL('sysdate'))
-    assert_equal "where sysdate = sysdate",       SQL.where(JDBCHelper::SQL('sysdate') => JDBCHelper::SQL('sysdate'))
+    assert_equal "where a = sysdate",             SQL.where(:a => SQL.expr('sysdate'))
+    assert_equal "where a = sysdate",             SQL.where(:a => SQL.expr('sysdate'))
+    assert_equal "where sysdate = sysdate",       SQL.where(SQL.expr('sysdate') => SQL.expr('sysdate'))
     assert_equal "where a in ('aa', 'bb', 'cc')", SQL.where(:a => %w[aa bb cc])
     assert_equal "where a in ('aa', 'bb', 'cc', 4)", SQL.where(:a => %w[aa bb cc] + [4])
     assert_equal "where a = 1 and b = 'A''s'",    SQL.where(:a => 1, :b => "A's")
     assert_equal "where (a = 1 or b = 1)",        SQL.where("a = 1 or b = 1")
-    assert_equal "where (a = 1 or b = 1)",        SQL.where(JDBCHelper::SQL("a = 1 or b = 1"))
+    assert_equal "where (a = 1 or b = 1)",        SQL.where(SQL.expr("a = 1 or b = 1"))
     assert_equal "where (a = 1 or b = 1) and c = 2", SQL.where("a = 1 or b = 1", :c => 2)
     assert_equal "where c = 2 and (a = 1 or b = 1)", SQL.where({:c => 2}, "a = 1 or b = 1")
     assert_equal "where c = 2 and (a = 1 or b = 1) and (e = 2) and f = 3 and (abc not like % || 'abc???' || % or def != 100 and ghi = '??') and (1 = 1)",
@@ -62,6 +84,34 @@ class TestSQL < Test::Unit::TestCase
                   ["abc not like % || ? || % or def != ? and ghi = '??'", 'abc???', 100], [], ['1 = 1'])
     assert_equal '', SQL.where(nil)
     assert_equal '', SQL.where(" ")
+
+    # 0.7.0
+    {
+      :gt => '>',
+      :lt => '<',
+      :ge => '>=',
+      :le => '<=',
+      :ne => '<>'
+    }.each do |x, op|
+      assert_equal "where a #{op} 5", SQL.where(:a => SQL.send(x, 5))
+      assert_equal "where a #{op} 'z'", SQL.where(:a => SQL.send(x, 'z'))
+    end
+    assert_equal "where a > sysdate", SQL.where(:a => SQL.gt(SQL.expr 'sysdate'))
+
+    # 0.7.0: like
+    assert_raise(ArgumentError) { SQL.like(123) }
+    assert_raise(ArgumentError) { SQL.not_like(BigDecimal.new('1.23')) }
+    assert_equal "where a like 'abc'",     SQL.where(:a => SQL.like('abc'))
+    assert_equal "where a not like 'abc'", SQL.where(:a => SQL.not_like('abc'))
+
+    # 0.7.0: with where
+    assert_equal "where a > 1",        SQL.where(:a => SQL.gt(1))
+    assert_equal "where a < 'z'",      SQL.where(:a => SQL.lt('z'))
+    assert_equal "where a >= sysdate", SQL.where(:a => SQL.ge(SQL.expr 'sysdate'))
+    assert_equal "where a <= sysdate", SQL.where(:a => SQL.le(SQL.expr 'sysdate'))
+    assert_equal "where a <> 'z'",     SQL.where(:a => SQL.ne(:z))
+    assert_equal "where a like 'hello''%'", SQL.where(:a => SQL.like("hello'%"))
+    assert_equal "where a not like 'hello''%'", SQL.where(:a => SQL.not_like("hello'%"))
 
     # Non-primitive datatypes not implemented (TODO?)
     assert_raise(NotImplementedError) { SQL.where(:a => Time.now) }
@@ -74,11 +124,11 @@ class TestSQL < Test::Unit::TestCase
     assert_raise(ArgumentError) { SQL.where(" 'a--b' -- cde", :a => 1) }
     assert_raise(ArgumentError) { SQL.where(" 'a--b' -- cde", :a => 1) }
     assert_raise(ArgumentError) { SQL.where({:a => 1}, "/* a") }
-    assert_raise(ArgumentError) { SQL.where(:a => JDBCHelper::SQL(" 'a--b' -- cde")) }
-    assert_raise(ArgumentError) { SQL.where(:a => JDBCHelper::SQL(" 'aabbb''dd")) }
-    assert_raise(ArgumentError) { SQL.where(:a => JDBCHelper::SQL(" 'aabbb''dd' /* aaa */")) }
-    assert_raise(ArgumentError) { SQL.where(:a => JDBCHelper::SQL(' aabbb"""  ')) }
-    assert_raise(ArgumentError) { SQL.where(:a => JDBCHelper::SQL(' aab`bb``  ')) }
+    assert_raise(ArgumentError) { SQL.where(:a => SQL.expr(" 'a--b' -- cde")) }
+    assert_raise(ArgumentError) { SQL.where(:a => SQL.expr(" 'aabbb''dd")) }
+    assert_raise(ArgumentError) { SQL.where(:a => SQL.expr(" 'aabbb''dd' /* aaa */")) }
+    assert_raise(ArgumentError) { SQL.where(:a => SQL.expr(' aabbb"""  ')) }
+    assert_raise(ArgumentError) { SQL.where(:a => SQL.expr(' aab`bb``  ')) }
   end
 
   def test_where_prepared
@@ -91,8 +141,9 @@ class TestSQL < Test::Unit::TestCase
     assert_equal ["where a is null", []],             SQLPrepared.where(:a => nil)
     assert_equal ["where a is not null", []],          SQLPrepared.where(:a => SQL.not_nil)
     assert_equal ["where a is not null", []],         SQLPrepared.where(:a => SQL.not_null)
-    assert_equal ["where a = sysdate", []],           SQLPrepared.where(:a => JDBCHelper::SQL('sysdate'))
-    assert_equal ["where sysdate = sysdate", []],     SQLPrepared.where(JDBCHelper::SQL('sysdate') => JDBCHelper::SQL('sysdate'))
+    assert_equal ["where a = sysdate", []],           SQLPrepared.where(:a => SQL.expr('sysdate'))
+    assert_equal ["where a = sysdate", []],           SQLPrepared.where(:a => SQL.expr('sysdate'))
+    assert_equal ["where sysdate = sysdate", []],     SQLPrepared.where(SQL.expr('sysdate') => SQL.expr('sysdate'))
     assert_equal ["where a in ('aa', 'bb', 'cc')", []], SQLPrepared.where(:a => %w[aa bb cc])
     assert_equal ["where a in ('aa', 'bb', 'cc', 4)", []], SQLPrepared.where(:a => %w[aa bb cc] + [4])
     assert_equal ["where a = ? and b = ?", [1, "A's"]],   SQLPrepared.where(:a => 1, :b => "A's")
@@ -106,6 +157,14 @@ class TestSQL < Test::Unit::TestCase
                   ["abc not like % || ? || % or def != ?", 'abc', 100], [], ['1 = 1'])
     assert_equal [nil, []], SQLPrepared.where(nil)
     assert_equal [nil, []], SQLPrepared.where(" ")
+
+    # 0.7.0: with where
+    assert_equal ["where a > ?", [1]],        SQLPrepared.where(:a => SQL.gt(1))
+    assert_equal ["where a < ?", ['z']],      SQLPrepared.where(:a => SQL.lt('z'))
+    assert_equal ["where a >= sysdate", []], SQLPrepared.where(:a => SQL.ge(SQL.expr 'sysdate'))
+    assert_equal ["where a <= sysdate", []], SQLPrepared.where(:a => SQL.le(SQL.expr 'sysdate'))
+    assert_equal ["where a like ?", ["hello'%"]], SQLPrepared.where(:a => SQL.like("hello'%"))
+    assert_equal ["where a not like ?", ["hello'%"]], SQLPrepared.where(:a => SQL.not_like("hello'%"))
   end
 
   def test_select
@@ -138,29 +197,30 @@ class TestSQL < Test::Unit::TestCase
 
   def test_update
     assert_equal "update a.b set a = 1, b = 'A''s', c = now()", 
-      SQL.update('a.b', {:a => 1, :b => "A's", :c => JDBCHelper::SQL('now()')}, {})
+      SQL.update('a.b', {:a => 1, :b => "A's", :c => SQL.expr('now()')}, {})
 
     assert_equal "update a.b set a = 1, b = 'A''s', c = now() where a is not null", 
-      SQL.update('a.b', {:a => 1, :b => "A's", :c => JDBCHelper::SQL('now()')}, { :a => SQL.not_nil })
+      SQL.update('a.b', {:a => 1, :b => "A's", :c => SQL.expr('now()')}, { :a => SQL.not_nil })
 
     assert_equal ["update a.b set a = ?, b = ?, c = now()", [1, "A's"]],
-      SQLPrepared.update('a.b', {:a => 1, :b => "A's", :c => JDBCHelper::SQL('now()')}, {})
+      SQLPrepared.update('a.b', {:a => 1, :b => "A's", :c => SQL.expr('now()')}, {})
 
     assert_equal ["update a.b set a = ?, b = ?, c = now() where a = ?", [1, "A's", 2]],
-      SQLPrepared.update('a.b', {:a => 1, :b => "A's", :c => JDBCHelper::SQL('now()')}, { :a => 2 })
+      SQLPrepared.update('a.b', {:a => 1, :b => "A's", :c => SQL.expr('now()')}, { :a => 2 })
   end
 
   def test_insert
     {'insert' => :insert, 'insert ignore' => :insert_ignore, 'replace' => :replace}.each do |op, met|
-      assert_equal "#{op} into a.b (a, b, c) values (1, 'A''s', null)",
-        SQL.send(met, 'a.b', :a => 1, :b => "A's", :c => nil)
+      assert_equal "#{op} into a.b (a, b, c, d) values (1, 'A''s', null, sysdate)",
+        SQL.send(met, 'a.b', :a => 1, :b => "A's", :c => nil, :d => SQL.expr('sysdate'))
 
-      assert_equal ["#{op} into a.b (a, b, c) values (?, ?, ?)", [1, "A's", nil]],
-        SQLPrepared.send(met, 'a.b', :a => 1, :b => "A's", :c => nil)
+      assert_equal ["#{op} into a.b (a, b, c, d) values (?, ?, ?, sysdate)", [1, "A's", nil]],
+        SQLPrepared.send(met, 'a.b', :a => 1, :b => "A's", :c => nil, :d => SQL.expr('sysdate'))
     end
   end
 
-  def test_sql_equality
+  def test_sql_expression_equality
+    # Backward-compatibility
     assert_equal "a = b", JDBCHelper.SQL('a = b').to_s
     assert_equal JDBCHelper.SQL('a = b'), JDBCHelper.SQL('a = b')
 
@@ -168,9 +228,16 @@ class TestSQL < Test::Unit::TestCase
     assert JDBCHelper.SQL('a = b') == (JDBCHelper.SQL('a = b'))
     assert JDBCHelper.SQL('a = b') == 'a = b'
     assert       JDBCHelper.SQL('a = b').eql?(JDBCHelper.SQL('a = b'))
+    assert_equal JDBCHelper.SQL('a = b').hash, JDBCHelper.SQL('a = b').hash
     assert_false JDBCHelper.SQL('a = b').eql?('a = b')
 
     assert JDBCHelper.SQL('a = b') != JDBCHelper.SQL('a = c')
+
+    h = {}
+    h[SQL.expr 'a = b'] = 1 
+    h[SQL.expr 'a = b'] = 2
+    assert_equal 1, h.keys.length
+    assert_equal 2, h[SQL.expr 'a = b']
   end
 end